fix(darwin): remove the need for cgo when building a darwin binary on linux
