+++ cmd/argocd-k8s-auth/commands/azure_cgo.go
@@ -1,3 +1,5 @@
+//go:build !darwin || (cgo && darwin)
+
 package commands
 
 import (
